Controller dropdown widget file rename
